+++ contracts/mocks/Incidents/ICMockCover.sol
@@ -63,7 +63,7 @@ contract ICMockCover {
         uint96(amount),
         uint32(block.timestamp + 1),
         uint32(period),
-        uint96(0)
+        uint16(0)
       ));
 
     coverId = covers.length - 1;
